+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 4.8.0-wip
+## 4.8.0
 
 * Add `Expression.operatorSubtract`
 * Deprecate `Expression.operatorSubstract`
@@ -29,6 +29,7 @@
 * Add `Expression.bitwiseOrAssign`
 * Allow passing an `Expression` through `literal` without an exception.
 * Add support for extension types.
+* Update SDK version constraints to `>=3.0.0`.
 
 ## 4.7.0